# Improve performance of the renderFile() function by up to 4x