These changes improve the performance of the `ejs.renderFile()` function by up to about 4x.
To be clear, the improvement is just in the `renderFile()` function, not the entire rendering process.

This increase in performance was achieved by:

+ Not calling `Array.prototype.slice` on the `arguments` object and then manipulating the returned array
+ Creating a global `_OPTS_EXPRESS` variable instead of manipulating the `_OPTS` array on every call to `renderFile()`
+ Moving the try-catch statement into its own function so that `renderFile` may be optimized by V8
+ Using a regular `for` loop in `utils.shallowCopyFromList()` instead of a `forEach` loop

### Benchmark

```js
'use strict';

const Benchmark = require('benchmark');

const ejs = require('ejs');
const ejs2 = require('/path/to/this/branch/ejs');

const filePath = './helloWorld.ejs'; // helloWorld.ejs just contains "Hello World!"
const renderData = {};
const renderOptions = {
  cache: true,
};
function cb() { }

new Benchmark.Suite()
  .add('before - 3 args', () => {
    ejs.renderFile(filePath, renderOptions, cb);
  })
  .add('after - 3 args', () => {
    ejs2.renderFile(filePath, renderOptions, cb);
  })
  .add('before - 4 args', () => {
    ejs.renderFile(filePath, renderData, renderOptions, cb);
  })
  .add('after - 4 args', () => {
    ejs2.renderFile(filePath, renderData, renderOptions, cb);
  })
  .on('cycle', (event) => {
    console.log(String(event.target));
  })
  .run();
```

Results (higher is better):

| # args | before (ops/sec) | after (ops/sec) | speedup factor
|--------|-------------------|------------------|-----------------
| 3         | 186,263              | 431,891            | 2.3
| 4         | 268,917              | 1,017,427          | 3.8

Raw results:

```
before - 3 args x 186,263 ops/sec ±0.82% (93 runs sampled)
after - 3 args x 431,891 ops/sec ±0.54% (97 runs sampled)
before - 4 args x 268,917 ops/sec ±0.45% (95 runs sampled)
after - 4 args x 1,017,427 ops/sec ±0.49% (91 runs sampled)
```