@nwoltman 2017-01-23
@RyanZim Thanks for approving the PR :)

Additionally, the [benchmark](https://www.npmjs.com/package/benchmark) module only outputs results in operations/second. I included the raw output of the benchmark in the PR description, although if you find it acceptable, here's the equivalent milliseconds/operation:

| # args | before        | after
|--------|--------------|---------
| 3         | 0.00537 ms | 0.00232 ms
| 4         | 0.00372 ms | 0.00098 ms