@mde 2017-01-25
I'm assuming these operations are all in Node -- is that correct? I imagine these differences are highly implementation-specific. (Might be worth doing this, as it's probably some amount faster in other runtimes as well.)