@nwoltman 2017-01-25
Yes the benchmark was conducted in Node, however, it is logical that the new code will perform better in other runtimes as well since it creates less short-lived objects that will need to be garbage-collected and minimizes memory manipulation.

The only thing that's specific to V8 is moving the try-catch to its own function, which doesn't impact other runtimes and hopefully doesn't add much complexity to the code.

The `renderFile()` function can only be used with Node anyway so it's not possible to run the benchmark against other runtimes ([Node.js on ChakraCore](https://github.com/nodejs/node-chakracore) is still a work in progress).