@mde 2017-01-25
Oh, duh, this is FS stuff. Yes, this all makes sense. Thanks for improving this!